Refactor ListPage
